requirements.txt
dgut-requests==0.1.0
chinesecalendar
click==7.1.2
pyYAML==5.4.1
